Refactor Backups ListView
