+++ Source/SelfService/Web/applications/backup/listView.tsx
@@ -3,17 +3,14 @@
 
 import React, { useState, useEffect } from 'react';
 
-import { useSnackbar } from 'notistack';
-
-import { DataGridPro, GridColDef } from '@mui/x-data-grid-pro';
-
-import { Paper, Typography } from '@mui/material';
-
-import { IconButton } from '@dolittle/design-system';
+import { Typography } from '@mui/material';
 
 import { HttpResponseApplication } from '../../apis/solutions/application';
-import { BackupLink, getLink, BackupsForApplication, getBackupsByApplication, BackupLinkShareInput } from '../../apis/solutions/backups';
+import { BackupsForApplication, getBackupsByApplication } from '../../apis/solutions/backups';
 
+import { BackupsDataGrid, BackupsDetailsList } from './backupsDataGrid';
+
+// Hack to get the date from the file name.
 const getDateFromFileName = (fileName: string) => {
     // Remove from string all characters that are not digits, underscores or dashes.
     // From 'petripoint-dev-2023-06-14_05-29-14' to '----2023-06-14_05-29-14'.
@@ -32,36 +29,12 @@ const getDateFromFileName = (fileName: string) => {
     return `${date} at ${time}`;
 };
 
-const getBackupShareLink = async (row: BackupsDetailsList, applicationId: string) => {
-    const input: BackupLinkShareInput = {
-        applicationId,
-        environment: row?.environment,
-        file_path: row?.file,
-    };
-
-    const share: BackupLink = await getLink(input);
-
-    return share.url;
-};
-
-type BackupsDetailsList = {
-    environment: string;
-    file: string;
-    createdOn: string;
-};
-
-type BackupsTableRows = {
-    row: BackupsDetailsList;
-};
-
 export type ListViewProps = {
     application: HttpResponseApplication;
     environment: string;
 };
 
 export const ListView = ({ application, environment }: ListViewProps) => {
-    const { enqueueSnackbar } = useSnackbar();
-
     const [data, setData] = useState({} as BackupsForApplication);
     const [isLoaded, setIsLoaded] = useState(false);
 
@@ -78,7 +51,7 @@ export const ListView = ({ application, environment }: ListViewProps) => {
     // TODO: Add loading indicator
     if (!isLoaded) return null;
 
-    const backupsDataGridRows: BackupsDetailsList[] = data.files.map<BackupsDetailsList>(file => {
+    const backups: BackupsDetailsList[] = data.files.map<BackupsDetailsList>(file => {
         return {
             id: file,
             file,
@@ -87,88 +60,11 @@ export const ListView = ({ application, environment }: ListViewProps) => {
         };
     });
 
-    const DownloadCell = ({ row }: BackupsTableRows) => {
-        const handleDownload = async () => {
-            const url = await getBackupShareLink(row, application.id);
-            window.open(url, '_blank');
-            enqueueSnackbar(`${row?.file} has been downloaded.`);
-        };
-
-        return (
-            <IconButton tooltipText='Download backup file' icon='DownloadRounded' onClick={handleDownload} />
-        );
-    };
-
-    const CopyLinkCell = ({ row }: BackupsTableRows) => {
-        const handleLinkCopy = async () => {
-            const url = await getBackupShareLink(row, application.id);
-            await navigator.clipboard.writeText(url);
-            enqueueSnackbar(`${row?.file} copied to clipboard.`);
-        };
-
-        return (
-            <IconButton tooltipText='Copy link' icon='CopyAllRounded' onClick={handleLinkCopy} />
-        );
-    };
-
-    const backupsDataGridColumns: GridColDef[] = [
-        {
-            field: 'file',
-            headerName: 'Name',
-            minWidth: 687,
-            flex: 1,
-        },
-        {
-            field: 'createdOn',
-            headerName: 'Date & Time',
-            headerAlign: 'center',
-            align: 'center',
-            minWidth: 277,
-            flex: 1,
-        },
-        {
-            field: 'download',
-            headerName: 'Download',
-            sortable: false,
-            headerAlign: 'center',
-            align: 'center',
-            minWidth: 104,
-            flex: 1,
-            renderCell: DownloadCell,
-        },
-        {
-            field: 'copy',
-            headerName: 'Copy',
-            sortable: false,
-            headerAlign: 'center',
-            align: 'center',
-            minWidth: 104,
-            flex: 1,
-            renderCell: CopyLinkCell,
-        },
-    ];
-
     return (
         <>
             <Typography variant='h1' sx={{ mt: 3 }}>{application.name}</Typography>
             <Typography variant='subtitle1' sx={{ my: 2 }}>{`${environment} Environment`}</Typography>
-
-            <Paper elevation={1} sx={{ width: 1, boxShadow: 'none' }}>
-                <DataGridPro
-                    rows={backupsDataGridRows}
-                    columns={backupsDataGridColumns}
-                    autoHeight
-                    headerHeight={46}
-                    getRowHeight={() => 'auto'}
-                    getEstimatedRowHeight={() => 40}
-                    disableColumnMenu
-                    disableColumnReorder
-                    disableColumnResize
-                    disableColumnSelector
-                    disableSelectionOnClick
-                    hideFooter
-                />
-            </Paper>
+            <BackupsDataGrid backupsDataGridRows={backups} applicationId={application.id} />
         </>
     );
 };
